+++ ParcelInvoiceMaker/ParcelInvoiceMaker/ParcelProcessor.swift
@@ -9,19 +9,11 @@ import Foundation
 class ParcelInformation {
     let receiverInfomation: Receiver.Infomation
     
-    let deliveryCost: Int
+    private let deliveryCost: Int
     private let discount: Discount
     var discountedCost: Int {
-        switch discount {
-        case .none:
-            return deliveryCost
-        case .vip:
-            return deliveryCost / 5 * 4
-        case .coupon:
-            return deliveryCost / 2
-        }
+        discount.strategy.applyDiscount(deliveryCost: deliveryCost)
     }
-
     
     init(receiverInfomation: Receiver.Infomation, deliveryCost: Int, discount: Discount) {
         self.receiverInfomation = receiverInfomation
@@ -31,9 +23,9 @@ class ParcelInformation {
 }
 
 
-enum Discount: Int {
-    case none = 0, vip, coupon
-}
+//enum Discount: Int {
+//    case none = 0, vip, coupon
+//}
 
 protocol OrderProcessor {
     func process(parcelInformation: ParcelInformation, onComplete: @escaping (ParcelInformation) -> Void)
